+++ Frameworks/OpenMPT/OpenMPT/src/mpt/base/constexpr_throw.hpp
@@ -8,6 +8,8 @@
 #include "mpt/base/detect.hpp"
 #include "mpt/base/namespace.hpp"
 
+#include <utility>
+
 
 
 namespace mpt {